[core] move SourceInfo to its own file
